get_module_doc move to mathics.doc.common_doc
