+++ setup.py
@@ -190,6 +190,7 @@ setup(
         "mathics.builtin.specialfns",
         "mathics.builtin.statistics",
         "mathics.builtin.string",
+        "mathics.builtin.systems_init",
         "mathics.builtin.vectors",
         "mathics.doc",
         "mathics.format",
